pyproject.toml
[project]
name = "mcp-obsidian"
version = "0.1.0"
description = "MCP Server for Obsidian vaults - read, write, search, organize and analyze your notes"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Erin Mikail Staples" },
]
keywords = ["mcp", "obsidian", "markdown", "notes", "goose"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Office/Business",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "python-frontmatter>=1.0.0",
    "ruamel.yaml>=0.18.0", 
    "send2trash>=1.8.0",
    "filelock>=3.12.0",
    "jmespath>=1.0.0",
    "chardet>=5.2.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
watch = ["watchdog>=4.0.0"]
fuzzy = ["rapidfuzz>=3.0.0"]
all = ["watchdog>=4.0.0", "rapidfuzz>=3.0.0"]

[project.scripts]
mcp-obsidian = "obsidian_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/erinmikailstaples/obsidian-mcp"
Repository = "https://github.com/erinmikailstaples/obsidian-mcp"
Issues = "https://github.com/erinmikailstaples/obsidian-mcp/issues"
Documentation = "https://github.com/erinmikailstaples/obsidian-mcp/blob/main/README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_mcp"]
